+++ nflow-explorer/src/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 import {readConfig, ConfigContext} from './config';
 import {createTheme, ThemeProvider} from '@mui/material/styles';
 import {TableCell} from '@mui/material';
@@ -81,8 +80,3 @@ readConfig().then(config => {
     </React.StrictMode>
   );
 });
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();